… lib/core/src/Cardano/Wallet/Primitive/AddressDerivation/Jormungandr.hs
… lib/core/test/unit/Cardano/Wallet/DB/SqliteSpec.hs
… lib/core/test/unit/Cardano/Wallet/Network/BlockHeadersSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDerivation/ByronSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDerivation/IcarusSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDerivation/JormungandrSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDerivationSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDiscovery/RandomSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDiscovery/SequentialSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDiscoverySpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/CoinSelection/LargestFirstSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/CoinSelection/MigrationSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/CoinSelection/RandomSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/CoinSelectionSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/FeeSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/ModelSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/SlottingSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/SyncProgressSpec.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/TypesSpec.hs
… lib/core/test/unit/Cardano/Wallet/RegistrySpec.hs
… lib/core/test/unit/Cardano/WalletSpec.hs
… lib/core/test/unit/Data/Function/UtilsSpec.hs
… lib/core/test/unit/Data/QuantitySpec.hs
… lib/core/test/unit/Data/Time/TextSpec.hs
… lib/core/test/unit/Data/Vector/ShuffleSpec.hs
… lib/core/test/unit/Network/Wai/Middleware/LoggingSpec.hs
… lib/test-utils/cardano-wallet-test-utils.cabal
… lib/test-utils/src/Test/QuickCheck/Extra.hs