# Issue Number

<!-- Put here a reference to the issue this PR relates to and which requirements it tackles -->

# Overview

<!-- Detail in a few bullet points the work accomplished in this PR -->

- 2a545b7f885a974ff64261538e83e6a250facd35
  :round_pushpin: **sprinkle some 'parallel' in unit tests**
    Some visible gain in time from a baseline of 362s down to 277s with
  -j4 & +RTS -N4. Unfortunately, we can't parallelise the database tests
  as easily.

- 44c88ad1326c7f12412ee581ca26427ee055b5e2
  :round_pushpin: **make db properties tests parallelisable and parallelise them.**
  
- c3f48d92240097ff9e7b5ab197bcea0fb873b48d
  :round_pushpin: **make pool db unit tests run faster (80s -> 16s)**
    I've mainly relaxed some of the coverage constraints (which are really only there to make sure that tests are going through enough interesting cases, but there's no need to expect a _strict_ coverage). I've also made sure to not generate lists that are _too large_ if not needed to not burn CPU for no reason.

- 0d15e194955cf3e3f4229dbc69758a3db0e2ba20
  :round_pushpin: **reduce required number of successes for some long properties**
    - We've seen the rebalance selection property fail in 1400 and 1997 tests, there's no need to go to 10000.

  - The timing test for the logging middleware could really be a unit test, we are using QuickCheck here just
    for the reporting and because we were too lazy to think about good arbitrary values. Testing with 10 values
    is already more than sufficient.

  - depleteUTxO is long, and we are testing for 5 different values, although some have very similar results. I've
    picked a more sparse scale to still get to test various scenarios, but reduce the overall property time.

  - Various properties for 'depleteUTxO' were run with max success of 1000, that's unecessary large.

- 60a6b1993ccc1f6c8648c36e49dc5ff669473aa8
  :round_pushpin: **remove 'JormungandrKey' and surroundings.**
    Fortunately, 'ShelleyKey' is a drop-in replacement for everything that referred to 'JormungandrKey' in the database. It's good to finally do this because, most of the tests were still executed on the 'JormungandrKey' whereas most of what's used in production is actually 'ShelleyKey' ... we know that the underlying implementation is equal, but that's a bit dodgy :| ...

# Comments

<!-- Additional comments or screenshots to attach if any -->

Note: this is eating up to 40% of my available RAM though, so about 10 GB. Seems still okay-ish ?

<!-- 
Don't forget to:

 ✓ Self-review your changes to make sure nothing unexpected slipped through
 ✓ Assign yourself to the PR
 ✓ Assign one or several reviewer(s)
 ✓ Once created, link this PR to its corresponding ticket
 ✓ Assign the PR to a corresponding milestone
 ✓ Acknowledge any changes required to the Wiki
-->